Preflight Script Update
* Updated welcome message to be more friendly
* Created functions to execute command checking and config checking
* Added check for Terraform binary
* Added warning about existing .terraform directory (see below)
* Reworked gcloud variable state checking (see below)
* Updated Terraform bucket name to include "cloudlamp"
* Preflight now calls `terraform init`

Existing Terraform Directory
----------------------------
If an existing .terraform directory is present, there is a chance
TF will attempt to reconcile state with a nonexistent project (or
components therein). Discarding the local file alleviates the need
to check these parameters. Useful in debugging.

Gcloud Variable Checking
------------------------
Upon testing in Cloud Shell, it turns out it emits a message which
begins with "Your active configuration", which was not present on
other CLI clients. Now we use wildcard matching to ensure the right
variables are set prior to execution.